refactor uri handler into its own file
