+++ package.json
@@ -53,7 +53,6 @@
         "command": "flutterflow.handleUri",
         "title": "Handle FlutterFlow URI"
       }
-
     ],
     "configuration": {
       "title": "FlutterFlow",
@@ -106,7 +105,14 @@
           "name": "FlutterFlow Warnings"
         }
       ]
-    }
+    },
+    "uriHandlers": [
+      {
+        "protocol": "vscode",
+        "uriPattern": "^vscode://flutterflow\\.custom-code-editor.*",
+        "command": "flutterflow.handleUri"
+      }
+    ]
   },
   "scripts": {
     "build": "./node_modules/typescript/bin/tsc",
