+++ src/extension.ts
@@ -16,6 +16,7 @@ import { performPullLatest } from "./actions/pullLatest";
 import { createEditStream, FFProjectState, ProjectState } from "./ffState/FFProjectState";
 import { FlutterFlowApiClient } from "./api/FlutterFlowApiClient";
 import { FlutterFlowMetadata, getInitialFile, setInitialFile } from "./ffState/FlutterFlowMetadata";
+import { handleFlutterFlowUri } from "./actions/uriHandler";
 
 // Pattern for watching custom code files
 const kCustomFilePattern = `**/{pubspec.yaml,lib/custom_code/**,lib/flutter_flow/custom_functions.dart}`;
@@ -298,136 +299,12 @@ export function activate(context: vscode.ExtensionContext): vscode.ExtensionCont
   context.subscriptions.push(
     vscode.window.registerUriHandler({
       handleUri(uri: vscode.Uri): vscode.ProviderResult<void> {
-        console.log('opening project from URI ', uri);
-        // Parse all parameters from query string
-        // Expected format: vscode://flutterflow.custom-code-editor?projectId={projectId}&branchName={branchId}&fileName={fileName}
-        const params = new URLSearchParams(uri.query);
-        const projectId = params.get('projectId');
-        if (!projectId) {
-          vscode.window.showErrorMessage('Invalid FlutterFlow URI format: missing projectId');
-          return;
+        try {
+          return handleFlutterFlowUri(uri, context);
+        } catch (err) {
+          console.error('Error handling FlutterFlow URI:', err);
+          vscode.window.showErrorMessage(`Error handling FlutterFlow URI: ${err}`);
         }
-
-        const branchName = params.get('branchName') || 'main';
-        const fileName = params.get('fileName') || '';
-
-        const openProject = async () => {
-          try {
-            // Check if API key is configured
-            const apiKey = getApiKey();
-            if (!apiKey) {
-              const setApiKey = await vscode.window.showInformationMessage(
-                'FlutterFlow API key not found. Would you like to set it now?',
-                { modal: true },
-                'Yes', 'No'
-              );
-              if (setApiKey === 'Yes') {
-                const key = await vscode.window.showInputBox({
-                  prompt: 'Enter your FlutterFlow API key',
-                  password: true,
-                  ignoreFocusOut: true
-                });
-                if (key) {
-                  await vscode.workspace.getConfiguration('flutterflow').update('userApiToken', key, true);
-                } else {
-                  return;
-                }
-              } else {
-                return;
-              }
-            }
-
-            // Check if download location is configured
-            const downloadLocation = vscode.workspace.getConfiguration('flutterflow').get<string>('downloadLocation');
-            if (!downloadLocation) {
-              const setLocation = await vscode.window.showInformationMessage(
-                'FlutterFlow download location not set. Would you like to set it now?',
-                { modal: true },
-                'Yes', 'No',
-              );
-              if (setLocation === 'Yes') {
-                const uris = await vscode.window.showOpenDialog({
-                  canSelectFiles: false,
-                  canSelectFolders: true,
-                  canSelectMany: false,
-                  title: 'Select Download Location'
-                });
-                if (uris && uris[0]) {
-                  await vscode.workspace.getConfiguration('flutterflow').update('downloadLocation', uris[0].fsPath, true);
-                } else {
-                  return;
-                }
-              } else {
-                return;
-              }
-            }
-
-
-            // Get download path from settings.
-            // TODO: should we assume this download path or prompt the user for it like the normal download flow?
-            const downloadPath = vscode.workspace.getConfiguration("flutterflow").get<string>("downloadLocation") || "";
-
-            //check if the download path is a valid directory
-            if (!fs.existsSync(downloadPath)) {
-              vscode.window.showErrorMessage(`Invalid download path. ${downloadPath} does not exist.`);
-              return;
-            }
-            // check if download path plus projectid exists
-            const projectDownloadPath = path.join(downloadPath, projectId);
-            const projectDownloadPathExists = fs.existsSync(projectDownloadPath);
-
-            const currentWorkspacePath = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath;
-
-            // If project is already open in current workspace
-            if (currentWorkspacePath === projectDownloadPath) {
-              if (fileName) {
-                const fullPath = path.join(projectDownloadPath, fileName);
-                if (fs.existsSync(fullPath)) {
-                  const doc = await vscode.workspace.openTextDocument(fullPath);
-                  await vscode.window.showTextDocument(doc);
-                }
-              }
-              // Ensure editor is initialized
-              await initCodeEditorFn();
-              return;
-            }
-
-            // Project exists but isn't open
-            if (projectDownloadPathExists) {
-              const choice = await vscode.window.showInformationMessage(
-                'Project already exists locally. What would you like to do?',
-                { modal: true },
-                'Open Existing',
-                'Download Fresh Copy'
-              );
-
-              if (choice === 'Open Existing') {
-                if (fileName) {
-                  await setInitialFile(projectDownloadPath, fileName);
-                }
-                await vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(projectDownloadPath));
-                return;
-              } else if (!choice) {
-                return; // User cancelled
-              }
-            }
-
-            // Download new copy
-            await downloadCodeWithPrompt(context, {
-              projectId,
-              branchName,
-              downloadLocation: downloadPath,
-              initialFile: fileName
-            });
-
-          } catch (err) {
-            console.error('handleUri:openProject error', err);
-            vscode.window.showErrorMessage(`Error opening project from URL: ${err}`);
-          }
-        };
-
-        // unawaited promise to open the project
-        return openProject();
       }
     })
   );
